Extract hashing to new utility
