+++ confetti/src/main/scala/com/thetradedesk/confetti/BehavioralConfigLoader.scala
@@ -1,23 +1,17 @@
 package com.thetradedesk.confetti
 
-import com.amazonaws.regions.Regions
-import com.amazonaws.services.s3.{AmazonS3, AmazonS3ClientBuilder, AmazonS3URI}
+import com.thetradedesk.confetti.utils.{HashUtils, S3Utils}
 import org.yaml.snakeyaml.Yaml
 
-import java.io.ByteArrayInputStream
-import java.nio.charset.StandardCharsets
-import java.security.MessageDigest
-import java.util.Base64
 import java.time.LocalDateTime
 import scala.collection.JavaConverters._
 
 class BehavioralConfigLoader(env: String, experimentName: Option[String], groupName: String, jobName: String) {
 
-  private val s3Client: AmazonS3 = AmazonS3ClientBuilder.standard().withRegion(Regions.US_EAST_1).build()
 
   def loadConfig(): Map[String, Any] = {
     val path = buildConfigPath()
-    val yamlStr = readFromS3(path)
+    val yamlStr = S3Utils.readFromS3(path)
     val yaml = new Yaml()
     val javaMap = yaml.load[java.util.Map[String, Any]](yamlStr)
     var config = javaMap.asScala.toMap[String, Any]
@@ -26,9 +20,9 @@ class BehavioralConfigLoader(env: String, experimentName: Option[String], groupN
     config += ("date_time" -> LocalDateTime.now().toString)
 
     val rendered = yaml.dump(config.asJava)
-    val hash = hashConfig(rendered)
+    val hash = HashUtils.sha256Base64(rendered)
     val runtimePath = s"s3://thetradedesk-mlplatform-us-east-1/configdata/confetti/runtime-configs/$env/$groupName/$jobName/$hash/behavioral_config.yml"
-    writeToS3(runtimePath, rendered)
+    S3Utils.writeToS3(runtimePath, rendered)
     config
   }
 
@@ -37,21 +31,4 @@ class BehavioralConfigLoader(env: String, experimentName: Option[String], groupN
     s"s3://thetradedesk-mlplatform-us-east-1/configdata/confetti/configs/$env/$expDir$groupName/$jobName/behavioral_config.yml"
   }
 
-  private def readFromS3(path: String): String = {
-    val uri = new AmazonS3URI(path)
-    val obj = s3Client.getObject(uri.getBucket, uri.getKey)
-    scala.io.Source.fromInputStream(obj.getObjectContent).mkString
-  }
-
-  private def writeToS3(path: String, data: String): Unit = {
-    val uri = new AmazonS3URI(path)
-    val bytes = data.getBytes(StandardCharsets.UTF_8)
-    val is = new ByteArrayInputStream(bytes)
-    s3Client.putObject(uri.getBucket, uri.getKey, is, null)
-  }
-
-  private def hashConfig(content: String): String = {
-    val digest = MessageDigest.getInstance("SHA-256").digest(content.getBytes(StandardCharsets.UTF_8))
-    Base64.getEncoder.encodeToString(digest)
-  }
 }
